build: move error/index so we don't fail to import
Ref: metriport/metriport-internal#1040
