+++ packages/api/src/external/carequality/document/process-document-query-results.ts
@@ -1,6 +1,6 @@
 import { S3Utils, createS3FileName } from "@metriport/core/external/aws/s3";
 import { out } from "@metriport/core/util/log";
-import { errorToString } from "@metriport/core/util/error/index";
+import { errorToString } from "@metriport/core/util/error/shared";
 import { capture } from "@metriport/core/util/notifications";
 import { DocumentReference } from "@metriport/ihe-gateway-sdk";
 import { DocumentReference as FHIRDocumentReference } from "@medplum/fhirtypes";
