+++ packages/api/src/external/carequality/document/query-documents.ts
@@ -1,5 +1,5 @@
 import { makeLambdaClient } from "@metriport/core/external/aws/lambda";
-import { errorToString } from "@metriport/core/util/error/index";
+import { errorToString } from "@metriport/core/util/error/shared";
 import { capture } from "@metriport/core/util/notifications";
 import { Patient } from "@metriport/core/domain/patient";
 import { Config } from "../../../shared/config";
