+++ packages/core/src/external/carequality/command/documents/send-doc-query-results.ts
@@ -5,7 +5,7 @@ import dayjs from "dayjs";
 import duration from "dayjs/plugin/duration";
 import { MetriportError } from "../../../../util/error/metriport-error";
 import { initSequelizeForLambda } from "../../../../util/sequelize";
-import { errorToString } from "../../../../util/error";
+import { errorToString } from "../../../../util/error/shared";
 import { DocumentQueryResult } from "../../document-query-result";
 
 const REQUEST_ID_COLUMN = "request_id";
