+++ packages/core/src/util/error/shared.ts
